OpenMP changes:
Use #pragma instead of _Pragma for OpenMP directives.  The PGI compilers do
not understand the _Pragma syntax.

Use int instead of ptrdiff_t for loop variables which are parallelised via
OpenMP.  The PGI compilers cannot handle ptrdiff_t.

+++ Carpet/CarpetLib/src/interpolate_eno_3d_3tl.cc
@@ -155,9 +155,9 @@ namespace CarpetLib {
     
     // Loop over region
 #pragma omp parallel for
-    for (ptrdiff_t k=0; k<regkext; ++k) {
-      for (ptrdiff_t j=0; j<regjext; ++j) {
-        for (ptrdiff_t i=0; i<regiext; ++i) {
+    for (int k=0; k<regkext; ++k) {
+      for (int j=0; j<regjext; ++j) {
+        for (int i=0; i<regiext; ++i) {
           
           T const s1 = src1 [SRCIND3(i, j, k)];
           T const s2 = src2 [SRCIND3(i, j, k)];
